updated response before running payslip

+++ Modules/Payslip/Http/Resources/SalaryItemsResource.php
@@ -36,7 +36,7 @@ class SalaryItemsResource extends JsonResource
         if($salary_item->salaryItemsName->name == "Wages"){
             if($salary_item->amount <= 0){
                 // $hours_worked = $this->get_hours_worked(request('employee_id'), request('from_date'), request('to_date'));
-                $hours_worked = (int)request('working_hours');
+                $hours_worked = (float)request('working_hours');
                 return round($hours_worked, 2) . " hours";
             }
             return "1 month";
@@ -44,31 +44,31 @@ class SalaryItemsResource extends JsonResource
         if($salary_item->salaryItemsName->salaryItemsCategory->id == 1)
         {
             if(request('maternity_leave') && $salary_item->salaryItemsName?->name == "Maternity Time Rate"){
-                return number_format((int)request('maternity_leave'), 2);
+                return number_format((float)request('maternity_leave'), 2);
             }
             if(request('annual_leave') && $salary_item->salaryItemsName?->name == "Holiday Rate"){
-                return number_format((int)request('annual_leave'), 2);
+                return number_format((float)request('annual_leave'), 2);
             }
             if(request('sick_leave') && $salary_item->salaryItemsName?->name == "Paid Sick Leave Rate"){
-                return number_format((int)request('sick_leave'), 2);
+                return number_format((float)request('sick_leave'), 2);
             }
             if(request('unpaid_sick_leave') && $salary_item->salaryItemsName?->name == "Unpaid Sick Leave Rate"){
-                return number_format((int)request('unpaid_sick_leave'), 2);
+                return number_format((float)request('unpaid_sick_leave'), 2);
             }
             if(request('absent') && $salary_item->salaryItemsName?->name == "Absent Rate"){
-                return number_format((int)request('absent'), 2);
+                return number_format((float)request('absent'), 2);
             }
             if(request('overtime') && $salary_item->salaryItemsName?->name == "Overtime Rate"){
-                return number_format((int)request('overtime'), 2);
+                return number_format((float)request('overtime'), 2);
             }
             if(request('double_overtime') && $salary_item->salaryItemsName?->name == "Double Overtime Rate"){
-                return number_format((int)request('double_overtime'), 2);
+                return number_format((float)request('double_overtime'), 2);
             }
             if(request('bonus') && $salary_item->salaryItemsName?->name == "Bonus"){
-                return number_format((int)request('bonus'), 2);
+                return number_format((float)request('bonus'), 2);
             }
             if(request('recuperated_hours') && $salary_item->salaryItemsName?->name == "Recuperated Hour"){
-                return number_format((int)request('recuperated_hours'), 2);
+                return number_format((float)request('recuperated_hours'), 2);
             }
             return 0; // this will come from no. of leave days
         }
@@ -141,31 +141,31 @@ class SalaryItemsResource extends JsonResource
         }
         if($salary_item->salaryItemsName->salaryItemsCategory->id == 1){
             if(request('maternity_leave') && $salary_item->salaryItemsName?->name == "Maternity Time Rate"){
-                return number_format((int)request('maternity_leave'), 2) * $this->getSalaryItemAmount("Maternity Time Rate")->amount;
+                return number_format((float)request('maternity_leave'), 2) * $this->getSalaryItemAmount("Maternity Time Rate")->amount;
             }
             if(request('annual_leave') && $salary_item->salaryItemsName?->name == "Holiday Rate"){
-                return number_format((int)request('annual_leave'), 2) * $this->getSalaryItemAmount("Holiday Rate")->amount;
+                return number_format((float)request('annual_leave'), 2) * $this->getSalaryItemAmount("Holiday Rate")->amount;
             }
             if(request('sick_leave') && $salary_item->salaryItemsName?->name == "Paid Sick Leave Rate"){
-                return number_format((int)request('sick_leave'), 2) * $this->getSalaryItemAmount("Paid Sick Leave Rate")->amount;
+                return number_format((float)request('sick_leave'), 2) * $this->getSalaryItemAmount("Paid Sick Leave Rate")->amount;
             }
             if(request('unpaid_sick_leave') && $salary_item->salaryItemsName?->name == "Unpaid Sick Leave Rate"){
-                return number_format((int)request('unpaid_sick_leave'), 2) * $this->getSalaryItemAmount("Unpaid Sick Leave Rate")->amount;
+                return number_format((float)request('unpaid_sick_leave'), 2) * $this->getSalaryItemAmount("Unpaid Sick Leave Rate")->amount;
             }
             if(request('absent') && $salary_item->salaryItemsName?->name == "Absent Rate"){
-                return number_format((int)request('absent'), 2) * $this->getSalaryItemAmount("Absent Rate")->amount;
+                return number_format((float)request('absent'), 2) * $this->getSalaryItemAmount("Absent Rate")->amount;
             }
             if(request('overtime') && $salary_item->salaryItemsName?->name == "Overtime Rate"){
-                return number_format((int)request('overtime'), 2) * $this->getSalaryItemAmount("Overtime Rate")->amount;
+                return number_format((float)request('overtime'), 2) * $this->getSalaryItemAmount("Overtime Rate")->amount;
             }
             if(request('double_overtime') && $salary_item->salaryItemsName?->name == "Double Overtime Rate"){
-                return number_format((int)request('double_overtime'), 2) * $this->getSalaryItemAmount("Double Overtime Rate")->amount;
+                return number_format((float)request('double_overtime'), 2) * $this->getSalaryItemAmount("Double Overtime Rate")->amount;
             }
             if(request('bonus') && $salary_item->salaryItemsName?->name == "Bonus"){
-                return number_format((int)request('bonus'), 2) * $this->getSalaryItemAmount("Bonus")->amount;
+                return number_format((float)request('bonus'), 2) * $this->getSalaryItemAmount("Bonus")->amount;
             }
             if(request('recuperated_hours') && $salary_item->salaryItemsName?->name == "Recuperated Hour"){
-                return number_format((int)request('recuperated_hours'), 2) * $this->getSalaryItemAmount("Recuperated Hour")->amount;
+                return number_format((float)request('recuperated_hours'), 2) * $this->getSalaryItemAmount("Recuperated Hour")->amount;
             }
             return 0;
         }
